Fix signal handling setup.
The signal handler was being assigned to the sa_handler member of
the sigaction struct, which is incorrect; it should be assigned to
the sa_sigaction member.  The sa_handler is only for using the default
handler or ignoring the signal.  This worked only because the two
fields are genrally a union, but new enough compilers will flag this
as an error.